1404: drop DirtyRangeTracker
